requirements.txt
aiogram
python-dotenv
sqlalchemy
asyncpg
alembic
httpx
opencv-python
pyzbar
numpy